backend/requirements.txt
Django==5.0.*
psycopg2-binary==2.9.*
djangorestframework==3.14.*
dj-database-url==2.1.*
gunicorn==21.2.*
python-dotenv==1.0.*
django-cors-headers==4.3.*
requests==2.31.*
celery==5.3.*
praw==7.7.*
stripe==7.0.*
django-filter==23.5.*
markdown==3.5.*
django-rest-auth==0.9.*
whitenoise==6.6.*
redis==5.0.*
pytest==7.4.*
pytest-django==4.6.*
coverage==7.3.*
isort==5.12.*
black==23.12.*